[NT-0] refac: Move DateTime header to includes
I would like to be able to include datetime in a public header file but am currently unable to do so. This change moves the datetime.h file to the public includes folder.
