+++ Library/src/Systems/Conversation/ConversationSystemHelpers.cpp
@@ -19,7 +19,7 @@
 #include "CSP/Common/StringFormat.h"
 #include "CSP/Multiplayer/Conversation/Conversation.h"
 #include "CSP/Systems/Assets/AssetCollection.h"
-#include "Common/DateTime.h"
+#include "CSP/Common/DateTime.h"
 #include "Debug/Logging.h"
 
 namespace csp::systems::ConversationSystemHelpers
