+++ Tests/src/InternalTests/SchedulerTests.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "Common/DateTime.h"
+#include "CSP/Common/DateTime.h"
 #include "Common/Scheduler.h"
 #include "TestHelpers.h"
