+++ Tests/src/PublicAPITests/ComponentTests/ConversationComponentTests.cpp
@@ -27,7 +27,7 @@
 #include "CSP/Multiplayer/SpaceEntity.h"
 #include "CSP/Systems/SystemsManager.h"
 #include "CSP/Systems/Users/UserSystem.h"
-#include "Common/DateTime.h"
+#include "CSP/Common/DateTime.h"
 #include "MultiplayerTestRunnerProcess.h"
 #include "RAIIMockLogger.h"
 #include "TestHelpers.h"
