+++ Tests/src/PublicAPITests/DateTimeTests.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "Common/DateTime.h"
+#include "CSP/Common/DateTime.h"
 #include "TestHelpers.h"
 
 #include "gtest/gtest.h"
