+++ Tests/src/PublicAPITests/UserSystemTests.cpp
@@ -22,7 +22,7 @@
 #include "CSP/Systems/SystemsManager.h"
 #include "CSP/Systems/Users/Profile.h"
 #include "CSP/Systems/Users/UserSystem.h"
-#include "Common/DateTime.h"
+#include "CSP/Common/DateTime.h"
 #include "Common/LoginStateData.h"
 #include "Common/Web/HttpPayload.h"
 #include "RAIIMockLogger.h"
